Added partial solution for part 4
I am adding my current progress on part 4.
I have implemented the Shannon-Fano, LZW algorithm. You can test it by
hardcoding some string in each code. Each program show the given string,
the encoded string, and the decoded string.

About the arithmetic coding, I think I am halfway done through it. My
program is able to encode a string, given this string and its
alphabet(or dictionary, as I call on the program). What the program does
not do yet is taking the string and generating the alphabet itself.
Generally for this program I would do a loop on the to-be-encoded-string
that checked how many times each character appeared to compute their
frequencies(that is what I did on the other algorithms, basically). The
problem is that on the course notes, on the arithmetic coding example,
the probabilities of the symbols there seem a little weird to me. I'll
try to implement on the same approach I did for the other algorithms and
see what happens.

Another issue is about LZW coding: to encode the strings, I use decimal
numbers for each character on the string. Thus, the encoded string would
be a string of decimal numbers(I used this approach because that's how
it is on the course notes). The problem is that I realized that the
professor asks for a bitstream on the project document, so I'll try to
figure a way to assign only binary numbers for the encoding, such that
it still works.

Also, I am not a 100% sure if it is OK to output a simple text file with
the binary encoded strings, or if we would have to really output a
binary file itself(which would be more difficult).